CCAR : INTERFACE CHANGE : add a boolean parameter regist with default value true to one Component constructor
It was needed to avoid calling registry two times for a python superv_component
Registry was called in the python constructor and in the corba proxy (class PySupervCompo)
that is used to implement DSC ports for python component.
As regist value is true by default there is no impact on other parts of KERNEL or of modules.