CCAR: 1- add a new type MachineDefinition in ContainerManager idl to separate machine definition parameters and machine parameters for requests to container manager (MachineParameters)
2- in Resource manager replace module by component. Selection of a resource is now
done with a component name and no more with a module name.
Now in CatalogResource.xml, we can specify for a resource that it supports
a list of components or modules.
Syntax for a component is:
<component name="xxxx" moduleName="yyy" />
Syntax for a module is:
<modules moduleName="yyyy"/>
When a module is added, a component with the same name is also added.