The default value of the working directory of a job is set to the one of the ressource used.
If neither the job nor the ressource defines a working directory, we use the previous rule
to define a working directory, based on $HOME.
There was a bug in the previous version because $HOME was not considered an absolute path.
A temporary fix is to use "/$HOME" instead of "$HOME".