Fix a bug causing concurrent access problems to the Resources Manager
Since the development of parallel container launch, the container
manager is multi-threaded, but it used directly the resource manager
which is not thread safe. With this fix, all calls from the container
manager to the resource manager are done through CORBA, and since the
resource manager runs in a single-thread POA, thread-safety is ensured.