[EDF-13129&13127&13128] omniNames process not killed.
When running "salome test" on "salomeintegration2" machine,
sometimes, several YACS tests hang on, because the process
"omniNames" is not killed. This commit fixes the issue.
Further work should be done in order to deal with concurrency
when writing to the file "*pidict" which contains the pids of the
servers created when launching salome.